refactor: :truck: move OrganizationStore and QueryParamsStore into RootStore
