+++ src/App/router/Root.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import { Outlet, useLocation, useNavigate } from 'react-router-dom';
 import { Layout } from 'components/';
 import { DEFAULT_ORG } from 'configs/api';
-import { useQueryParamsStoreInit } from 'store/QueryParamsStore';
+import { useQueryParamsStoreInit } from 'store/';
 
 const Root = observer(() => {
   const navigate = useNavigate();
