+++ src/store/RootStore/QueryParamsStore/index.ts
@@ -1,2 +1,1 @@
 export { default } from './QueryParamsStore';
-export * from './hooks/useQueryParamsInit';
