+++ src/store/RootStore/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useRootStore';
+export * from './useQueryParamsInit';
