+++ src/store/RootStore/index.ts
@@ -1,3 +1,5 @@
 export { default as RootStoreProvider } from './RootStoreProvider';
 export { default as RootStore } from './RootStore';
+export { default as QueryParamsStore } from './QueryParamsStore';
+export { default as OrganizationStore } from './OrganizationStore';
 export * from './hooks';
