+++ src/store/index.ts
@@ -1,6 +1,3 @@
 export * from './RootStore';
-export * from './QueryParamsStore';
 export * from './models';
-// export * from './types';
-export { default as OrganizationStore } from './OrganizationStore';
 export { default as RepoPageStore } from './RepoPageStore';
